[Fix] 5.1.1/0.2.0/SH 로그인 이후 화면 에러 수정
