+++ Changelog.md
@@ -36,3 +36,7 @@
 - [Fix] 5.10.2/0.1.3/MS Merge 해결
 - [add] 5.10.2/0.1.4/WY  수정페이지 이동 기능 추가
 
+2024-03-25
+- [Feat] 5.1.1/0.1.0/SH-SM 로그인 화면 추가
+- [Fix] 5.1.1/0.2.0/SH 로그인 기능 수정
+
