+++ src/views/Home.vue
@@ -1,0 +1,85 @@
+<template>
+    <header>
+      <img alt="Vue logo" class="logo" src="@/assets/logo.svg" width="125" height="125" />
+      <div class="wrapper">
+        <HelloWorld msg="You did it!" />
+        <nav>
+          <RouterLink to="/viewproduct">판매상품 조회</RouterLink>
+          <RouterLink to="/employeeslist">사원리스트 조회</RouterLink>
+          <RouterLink to="/ppcview">거래처별 판매상품 리스트 조회</RouterLink>
+          <RouterLink to="/ppccrud">거래처별 판매상품 리스트 등록</RouterLink>
+          <RouterLink to="/clientsave">매출 거래처 등록</RouterLink>
+          <RouterLink to="/viewclient">매출 거래처 조회</RouterLink>
+        </nav>
+      </div>
+    </header>
+    <RouterView />
+  </template>
+  
+  <script setup>
+  import { RouterLink, RouterView } from 'vue-router'
+  </script>
+
+<style scoped>
+header {
+  line-height: 1.5;
+  max-height: 100vh;
+}
+
+.logo {
+  display: block;
+  margin: 0 auto 2rem;
+}
+
+nav {
+  width: 100%;
+  font-size: 12px;
+  text-align: center;
+  margin-top: 2rem;
+}
+
+nav a.router-link-exact-active {
+  color: var(--color-text);
+}
+
+nav a.router-link-exact-active:hover {
+  background-color: transparent;
+}
+
+nav a {
+  display: inline-block;
+  padding: 0 1rem;
+  border-left: 1px solid var(--color-border);
+}
+
+nav a:first-of-type {
+  border: 0;
+}
+
+@media (min-width: 1024px) {
+  header {
+    display: flex;
+    place-items: center;
+    padding-right: calc(var(--section-gap) / 2);
+  }
+
+  .logo {
+    margin: 0 2rem 0 0;
+  }
+
+  header .wrapper {
+    display: flex;
+    place-items: flex-start;
+    flex-wrap: wrap;
+  }
+
+  nav {
+    text-align: left;
+    margin-left: -1rem;
+    font-size: 1rem;
+
+    padding: 1rem 0;
+    margin-top: 1rem;
+  }
+}
+</style>
